+++ R/reexports.R
@@ -1,0 +1,10 @@
+# This file contains several re-export constants from other packages.
+# The approach follows the suggestion found at
+# https://stackoverflow.com/questions/52690307/reexporting-datasets-in-r-packages
+
+
+#' @inherit IEATools::nonenergy_products description source references title
+#' @examples
+#' Recca::nonenergy_products
+#' @export
+nonenergy_products <- IEATools::nonenergy_products