+++ tests/testthat/test-reexport.R
@@ -1,0 +1,4 @@
+test_that("re-exoprting works correctly", {
+  exists("nonenergy_products", where = "package:Recca", mode= "function")
+  Recca::nonenergy_products
+})