+++ NEWS.md
@@ -8,7 +8,7 @@ Cite all releases with doi [10.5281/zenodo.5226085](https://doi.org/10.5281/zeno
 which always resolves to the latest release.
 
 
-# Recca 0.1.47 (2023-10-23) 
+# Recca 0.1.47 (2023-10-23) [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10034476.svg)](https://doi.org/10.5281/zenodo.10034476)
 
 * Fixed a bug where EIOU efficiency calculated by calc_eta_fu_Y_eiou()
   gave the **0** matrix.